Make config files more idempotent
This change ensures that the config files are written after the mysql
package is installed, and any existing config files are simply
overwritten instead of being backed up.  This is needed so that when
purge is turned on for the config directory that no unnecessary files
are created by puppet itself, or package installation making backups.